requirements.txt
-i https://pypi.org/simple
certifi==2025.4.26; python_version >= '3.6'
charset-normalizer==3.4.2; python_version >= '3.7'
dnspython==2.7.0; python_version >= '3.9'
idna==3.10; python_version >= '3.6'
oauthlib==3.2.2; python_version >= '3.6'
pymongo==4.13.0; python_version >= '3.9'
python-dotenv==1.1.0; python_version >= '3.9'
requests==2.32.4; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
tweepy==4.15.0; python_version >= '3.9'
urllib3==2.4.0; python_version >= '3.9'
